test(5419): extract the duplicated spec helpers into one module
withLedgerTx, withSealedLedgerTx, setRequiredSigners and checkSubsetOf were
byte-identical in TransactionLedgerSpec and TransactionSpec, so every change to
one of them was a change owed to two files. ShowOrd, which only checkSubsetOf
uses, was duplicated the same way and moves with it.

They now live in Cardano.Wallet.Shelley.TransactionSpecSupport. The module name
does not end in Spec, so hspec-discover does not pick it up as a suite.

This is inside REQ-5 as clarified in the previous commit: the fence's subject
is production code, and no path under lib/*/src/ or lib/*/lib/ is touched.
