+++ lib/unit/cardano-wallet-unit.cabal
@@ -262,6 +262,7 @@ test-suite unit
     Cardano.Wallet.Shelley.NetworkSpec
     Cardano.Wallet.Shelley.TransactionLedgerSpec
     Cardano.Wallet.Shelley.TransactionSpec
+    Cardano.Wallet.Shelley.TransactionSpecSupport
     Cardano.Wallet.Submissions.Gen
     Cardano.Wallet.Submissions.OperationsSpec
     Cardano.Wallet.Submissions.PrimitivesSpec
