+++ lib/unit/test/unit/Cardano/Wallet/Shelley/TransactionLedgerSpec.hs
@@ -111,11 +111,9 @@ import Cardano.Ledger.Api
     )
 import Cardano.Ledger.Api.Tx.Body
     ( collateralInputsTxBodyL
-    , guardsTxBodyL
     , inputsTxBodyL
     , mintTxBodyL
     , outputsTxBodyL
-    , reqSignerHashesTxBodyL
     , withdrawalsTxBodyL
     )
 import Cardano.Ledger.BaseTypes
@@ -315,6 +313,12 @@ import Cardano.Wallet.Shelley.Transaction.Ledger
     , sealWriteTx
     , txConstraints
     )
+import Cardano.Wallet.Shelley.TransactionSpecSupport
+    ( checkSubsetOf
+    , setRequiredSigners
+    , withLedgerTx
+    , withSealedLedgerTx
+    )
 import Cardano.Wallet.Transaction
     ( DelegationAction (..)
     , PreSelection (..)
@@ -384,9 +388,6 @@ import Data.Maybe
     ( fromJust
     , isJust
     , maybeToList
-    )
-import Data.Ord
-    ( comparing
     )
 import Data.Proxy
     ( Proxy (..)
@@ -475,7 +476,6 @@ import Prelude
 import qualified Cardano.Api as Cardano
 import qualified Cardano.Balance.Tx.Eras as Write
     ( Conway
-    , IsRecentEra
     , RecentEra (RecentEraConway, RecentEraDijkstra)
     )
 import qualified Cardano.Balance.Tx.Primitive as BT
@@ -512,7 +512,6 @@ import qualified Data.Map.Strict as Map
 import qualified Data.Set as Set
 import qualified Data.Text as T
 import qualified Data.Text.Encoding as T
-import qualified GHC.Exts as Exts
 
 spec :: Spec
 spec = describe "TransactionSpec" $ do
@@ -807,81 +806,6 @@ lookupFnFromKeys keys addr =
                     (first (liftRawKey ShelleyKeyS) <$> keys)
     in
         Map.lookup addr addrMap
-
-withLedgerTx
-    :: Write.IsRecentEra era
-    => RecentEra era
-    -> (Write.Tx era -> Write.Tx era)
-    -> (Write.Tx era -> Property)
-    -> Property
-withLedgerTx recentEra modifyTx cont =
-    cardanoApiEraConstraints recentEra
-        $ forAllBlind
-            ( fromCardanoApiTx
-                <$> genTxInEra (cardanoEraFromRecentEra recentEra)
-            )
-            (cont . modifyTx)
-
-withSealedLedgerTx
-    :: RecentEra era
-    -> SealedTx
-    -> (Write.Tx era -> a)
-    -> a
-withSealedLedgerTx recentEra sealedTx cont = case recentEra of
-    Write.RecentEraConway -> case unsafeReadTx sealedTx of
-        Read.EraValue (Read.Tx tx :: Read.Tx txEra) -> case Read.theEra @txEra of
-            Read.Conway -> cont tx
-            _ -> eraMismatch
-    Write.RecentEraDijkstra -> case unsafeReadTx sealedTx of
-        Read.EraValue (Read.Tx tx :: Read.Tx txEra) -> case Read.theEra @txEra of
-            Read.Dijkstra -> cont tx
-            _ -> eraMismatch
-  where
-    eraMismatch = error "withSealedLedgerTx: transaction era mismatch"
-
-setRequiredSigners
-    :: RecentEra era
-    -> Set.Set (LedgerKeys.KeyHash LedgerKeys.Guard)
-    -> Write.Tx era
-    -> Write.Tx era
-setRequiredSigners recentEra requiredSignerHashes = case recentEra of
-    Write.RecentEraConway ->
-        bodyTxL . reqSignerHashesTxBodyL .~ requiredSignerHashes
-    -- TODO(#5209): the Dijkstra arm below is unverified, and nothing on this
-    -- branch can verify it -- the era only activates at the hard fork. Dijkstra
-    -- drops @reqSignerHashesTxBodyL@ (it is @notSupportedInThisEraL@ there) and
-    -- reaches required signers only through @reqSignerHashesTxBodyG@, a getter
-    -- over @guardsTxBodyL@ that keeps @KeyHashObj@ credentials and discards
-    -- @ScriptHashObj@ ones; writing @KeyHashObj@ credentials into guards is the
-    -- inverse of that getter, which is why it is plausible, not why it is
-    -- correct. It is also invisible to the Dijkstra census, whose population is
-    -- code that announces itself unimplemented: an @error@ or a @pendingWith@
-    -- whose message names the era. A stub fails loudly; this one succeeds, and
-    -- would succeed wrongly.
-    Write.RecentEraDijkstra ->
-        bodyTxL . guardsTxBodyL
-            .~ Exts.fromList (SL.KeyHashObj <$> Set.toList requiredSignerHashes)
-
-checkSubsetOf :: (Eq a, Show a) => [a] -> [a] -> Property
-checkSubsetOf as bs =
-    property
-        $ counterexample counterexampleText
-        $ all ((`Set.member` ys) . ShowOrd) as
-  where
-    xs = Set.fromList (ShowOrd <$> as)
-    ys = Set.fromList (ShowOrd <$> bs)
-
-    counterexampleText =
-        unlines
-            [ "the following set:"
-            , showSet xs
-            , "is not a subset of:"
-            , showSet ys
-            , "rogue elements:"
-            , showSet (xs `Set.difference` ys)
-            ]
-      where
-        showSet = pretty . fmap (show . unShowOrd) . F.toList
 
 prop_signTransaction_addsTxInWitnesses
     :: AnyRecentEra
@@ -2996,18 +2920,6 @@ instance Buildable a => Show (ShowBuildable a) where
 
 instance Arbitrary (Hash "Datum") where
     arbitrary = pure $ Hash $ BS.pack $ replicate 28 0
-
---------------------------------------------------------------------------------
--- Utilities
---------------------------------------------------------------------------------
-
--- | A convenient wrapper type that allows values of any type with a 'Show'
---   instance to be ordered.
-newtype ShowOrd a = ShowOrd {unShowOrd :: a}
-    deriving (Eq, Show)
-
-instance (Eq a, Show a) => Ord (ShowOrd a) where
-    compare = comparing show
 
 --------------------------------------------------------------------------------
 -- V2 key witness parity properties
